Update sessions_controller.rb
If we sessions clear here we will empty the shopping cart as well.  From an e-commerce point of view that is definitely something that you don't want to do with a user log out.  let them log out and keep their cart with the items they may have in them so when they re-visit they may be tempted to then continue buying the products they walked away.  Actually works on me with amazon.